requirements.txt
langchain
langchain-ollama
langchain-openai
langchain-core
langchain-community
python-dotenv
streamlit
flask
markdown